alx_travel_app/requirement.txt
django
djangorestframework
django-cors-headers
drf-yasg
django-environ
celery
mysqlclient
rabbitmq-server
requests 